fix: address metrics PR feedback
