+++ package.json
@@ -53,7 +53,6 @@
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
     "crypto-js": "^4.2.0",
-    "date-fns": "^4.4.0",
     "framer-motion": "^12.0.0",
     "immer": "^11.0.0",
     "jspdf": "^4.0.0",
